feat: better naming conventions
BREAKING CHANGE: change names of timing and interpolation functions

`TweenFunction` is now called `InterpolationFunction`.
Individual functions are now called `[type]Lerp` instead of `[type]Tween`.
For instance: `colorTween` is now `colorLerp`.

`InterpolationFunction` is now called `TimingFunction`.
This name is better aligned with the CSS spec.
