+++ packages/core/src/tweening/index.ts
@@ -1,4 +1,4 @@
 export * from './Animator';
 export * from './interpolationFunctions';
+export * from './timingFunctions';
 export * from './tween';
-export * from './tweenFunctions';
