+++ packages/docs/docs/guides/migration/12.0.0.md
@@ -104,3 +104,13 @@ Similarly, change your types reference in `motion-canvas.d.ts`:
 - /// <reference types="@motion-canvas/core/project" />
 + /// <reference types="@motion-canvas/legacy/project" />
 ```
+
+## Update function names
+
+`TweenFunction`s are now called `InterpolationFunction`s.
+The naming convention has changed. Instead of `[type]Tween` they now use
+`[type]Lerp`. For instance: `colorTween` is now `colorLerp`.
+
+`InterpolationFunction`s are now called `TimingFunction`s.
+Their individual names haven't changed (`easeInExpo`, `easeOutCubic`, etc.),
+but referring to them as timing functions is better aligned with the CSS spec.
